Attempt to handle the Brawl cutover better
Still not great. The problem is that `generatePin` was calculating the
next brawl id start time... which wasn't actually the problem. The
_real_ issue was that if I want to update the `current.json` file in
advance then it would screw up the id of the next "start" brawl pin.

e.g. The next brawl will be 40. I don't know the details yet so
`current.json` still had the details of Brawl 39. When the cutover
happens (4am here) there will be a period of time where the app and
Lambda jobs will be trying to use the details from current and it would
be wrong. I'd rather update it with some "please wait" text and get it
ready to update as soon as I can with the real details without it
affecting the "brawl end" pin. So if I update it to `id: 40` it means
the pin generating logic would use `brawl-end-41` as an id which is
wrong.

I think the real solution will involve getting rid of `current.json`
while I don't know the real details... needs more thought.